Update README.md
Data-Science-Assignment-No.1-Basic-Statics-Level-1
Q1) Identify the Data type for the Following:
Activity	Data Type
Number of beatings from Wife	
Results of rolling a dice	
Weight of a person	
Weight of Gold	
Distance between two places	
Length of a leaf	
Dog's weight	
Blue Color	
Number of kids	
Number of tickets in Indian railways	
Number of times married	
Gender (Male or Female)	


Q2) Identify the Data types, which were among the following
Nominal, Ordinal, Interval, Ratio.
Data	Data Type
Gender	
High School Class Ranking	
Celsius Temperature	
Weight	
Hair Color	
Socioeconomic Status	
Fahrenheit Temperature	
Height	
Type of living accommodation	
Level of Agreement	
IQ(Intelligence Scale)	
Sales Figures	
Blood Group	
Time Of Day	
Time on a Clock with Hands	
Number of Children	
Religious Preference	
Barometer Pressure	
SAT Scores	
Years of Education	


Q3) Three Coins are tossed, find the probability that two heads and one tail are obtained?

Q4)  Two Dice are rolled, find the probability that sum is
a)	Equal to 1
b)	Less than or equal to 4
c)	Sum is divisible by 2 and  3


Q5)  A bag contains 2 red, 3 green and 2 blue balls. Two balls are drawn at random. What is the probability that none of the balls drawn is blue?


Q6) Calculate the Expected number of candies for a randomly selected child 
Below are the probabilities of count of candies for children (ignoring the nature of the child-Generalized view)
CHILD	Candies count	Probability
A	1	0.015
B	4	0.20
C	3	0.65
D	5	0.005
E	6	0.01
F	2	0.120
Child A – probability of having 1 candy = 0.015.
Child B – probability of having 4 candies = 0.20

Q8) Calculate Expected Value for the problem below
a)	The weights (X) of patients at a clinic (in pounds), are
108, 110, 123, 134, 135, 145, 167, 187, 199
Assume one of the patients is chosen at random. What is the Expected Value of the Weight of that patient?

Q10) Draw inferences about the following boxplot & histogram

 Q13) What is the nature of skewness when mean, median of data are equal?
Q14) What is the nature of skewness when mean > median ?
Q15) What is the nature of skewness when median > mean?
Q16) What does positive kurtosis value indicates for a data ?
Q17) What does negative kurtosis value indicates for a data?
Q18) Answer the below questions using the below boxplot visualization.
What can we say about the distribution of the data?
What is nature of skewness of the data?
What will be the IQR of the data (approximately)? 


Q19) Comment on the below Boxplot visualizations? 
 
Draw an Inference from the distribution of data for Boxplot 1 with respect Boxplot 2.